Document bug in create-element-to-jsx, add Contributing section
As mentioned in https://github.com/reactjs/react-codemod/issues/59 , the
`create-element-to-jsx` codemod can destroy comments, so leave a warning in the
docs so people know to be careful.

Also add a "Support and Contributing" section at the bottom of the README that
briefly describes the state and purpose of the repo and how contributions are
handled.