Fix crash due to skipped children in create-element-to-jsx
In #79, I made it so the transformation is skipped in some cases, in particular
when the capitalization is invalid. This broke an assumption elsewhere in the
code that `convertNodeToJSX` always returns a `JSXElement`, so there was a crash
if a skipped element was used as a child. To fix, we can just wrap in a
`JSXExpressionContainer` for that case.