Fix incorrect prop value strings in create-element-to-jsx (#82)
JSX prop value strings are treated more like HTML strings than JavaScript
strings, so things like escape sequences don't work: it just uses a literal
backslash character. Also, if the string contains a `"` character, recast will
generate a double-quoted string with the `"` characters escaped, which isn't
valid for JSX prop strings. To fix, we wrap the string in curly braces if it
directly uses an escape sequence (i.e. the underlying code has a backslash) or
if it has a `"` character.