Updated so default order to include React ^16.3 lifecylcle methods.
Updated so that the default order includes the new React lifecycle methods.  Also, according to the latest eslint rules for React 16.3 and higher, static methods no longer always have to be at the top.  It depends on the name of the static method.  Also, one of the new lifecycle methods, getDerivedStateFromProps, is a static method that must go under the constructor.  Static methods that start with get, set, must go under the lifecycle methods.